backned/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
